requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
build==1.3.0
    # via pip-tools
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via
    #   reportlab
    #   requests
click==8.3.1
    # via pip-tools
coverage[toml]==7.12.0
    # via pytest-cov
distro==1.9.0
    # via openai
dropbox==12.0.2
    # via -r requirements.in
filelock==3.20.0
    # via -r requirements.in
google-api-core==2.28.1
    # via google-api-python-client
google-api-python-client==2.187.0
    # via -r requirements.in
google-auth==2.41.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.1
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.3
    # via -r requirements.in
googleapis-common-protos==1.72.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
jiter==0.12.0
    # via openai
oauthlib==3.3.1
    # via requests-oauthlib
openai==2.8.1
    # via -r requirements.in
packaging==25.0
    # via
    #   build
    #   pytest
pdf2image==1.17.0
    # via -r requirements.in
pillow==12.0.0
    # via
    #   pdf2image
    #   reportlab
pip-tools==7.5.2
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
ply==3.11
    # via stone
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pyparsing==3.2.5
    # via httplib2
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.1
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==7.0.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
reportlab==4.4.5
    # via -r requirements.in
requests==2.32.5
    # via
    #   dropbox
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
ruff==0.14.7
    # via -r requirements.in
six==1.17.0
    # via
    #   dropbox
    #   stone
sniffio==1.3.1
    # via openai
stone==3.3.1
    # via dropbox
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
